+++ tests/api/cache-control.test.ts
@@ -60,6 +60,7 @@ describe('cache-control tests', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
       tx_count: 1,
+      tx_total_size: 1,
       signer_bitvec: null,
       signer_signatures: null,
     };