+++ tests/api/mempool.test.ts
@@ -573,6 +573,7 @@ describe('mempool tests', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
       tx_count: 1,
+      tx_total_size: 1,
       signer_bitvec: null,
       signer_signatures: null,
     };
@@ -1355,6 +1356,7 @@ describe('mempool tests', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
       tx_count: 1,
+      tx_total_size: 1,
       signer_bitvec: null,
       signer_signatures: null,
     };
@@ -1432,6 +1434,7 @@ describe('mempool tests', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
       tx_count: 1,
+      tx_total_size: 1,
       signer_bitvec: null,
       signer_signatures: null,
     };
@@ -1661,6 +1664,7 @@ describe('mempool tests', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
       tx_count: 1,
+      tx_total_size: 1,
       signer_bitvec: null,
       signer_signatures: null,
     };
@@ -1685,6 +1689,7 @@ describe('mempool tests', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
       tx_count: 1,
+      tx_total_size: 1,
       signer_bitvec: null,
       signer_signatures: null,
     };
@@ -1812,6 +1817,7 @@ describe('mempool tests', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
       tx_count: 1,
+      tx_total_size: 1,
       signer_bitvec: null,
       signer_signatures: null,
     };
@@ -1836,6 +1842,7 @@ describe('mempool tests', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
       tx_count: 1,
+      tx_total_size: 1,
       signer_bitvec: null,
       signer_signatures: null,
     };
@@ -1860,6 +1867,7 @@ describe('mempool tests', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
       tx_count: 1,
+      tx_total_size: 1,
       signer_bitvec: null,
       signer_signatures: null,
     };