+++ tests/api/other.test.ts
@@ -62,6 +62,7 @@ describe('other tests', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
       tx_count: 1,
+      tx_total_size: 1,
       signer_bitvec: null,
       signer_signatures: null,
     };