+++ tests/api/search.test.ts
@@ -64,6 +64,7 @@ describe('search tests', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
       tx_count: 1,
+      tx_total_size: 1,
       signer_bitvec: null,
       signer_signatures: null,
     };
@@ -278,6 +279,7 @@ describe('search tests', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
       tx_count: 1,
+      tx_total_size: 1,
       signer_bitvec: null,
       signer_signatures: null,
     };
@@ -620,6 +622,7 @@ describe('search tests', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
       tx_count: 1,
+      tx_total_size: 1,
       signer_bitvec: null,
       signer_signatures: null,
     };
@@ -1067,6 +1070,7 @@ describe('search tests', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
       tx_count: 1,
+      tx_total_size: 1,
       signer_bitvec: null,
       signer_signatures: null,
     };